backend/pyproject.toml
[project]
name = "accountant-backend"
version = "0.1.0"
description = "Document vault & accounting suite - backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.19",
    "python-magic-bin>=0.4.14",
    "Pillow>=11.0.0",
    "pdf2image>=1.17.0",
    "httpx>=0.28.0",
    "anthropic>=0.42.0",
    "websockets>=14.0",
    "email-validator>=2.0.0",
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
    "python-dateutil>=2.9.0",
    "cryptography>=42.0.0",
    "APScheduler>=3.10.0",
    "aiosmtplib>=3.0.0",
    "Jinja2>=3.1.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth>=2.25.0",
    "plaid-python>=22.0.0",
    "stripe>=7.0.0",
    "twilio>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "pytest-cov>=6.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
